🎨 Lots of cleanup
